move HypixelJoinEvent
